+++ TMessagesProj/src/main/java/org/telegram/ui/GroupCallTabletGridAdapter.java
@@ -130,7 +130,7 @@ public class GroupCallTabletGridAdapter extends RecyclerListView.SelectionAdapte
         }
     }
 
-    public void scrollToPeerId(int peerId, RecyclerListView fullscreenUsersListView) {
+    public void scrollToPeerId(long peerId, RecyclerListView fullscreenUsersListView) {
 //        for (int i = 0; i < participants.size(); i++) {
 //            if (peerId == MessageObject.getPeerId(participants.get(i).peer)) {
 //                ((LinearLayoutManager) fullscreenUsersListView.getLayoutManager()).scrollToPositionWithOffset(i, AndroidUtilities.dp(13));
